tos-img-resize-python/requirements.txt
pillow
tos